refactor: ConsentsCom... > CommunicationCard
