+++ src/client/components/CommunicationCard.tsx
